+++ src/components/NavBar/NavBar.jsx
@@ -8,15 +8,15 @@ const NavBar = ({ user, handleLogout }) => {
       {user ?
         <ul>
           <li>Welcome, {user.name}</li>
-          <li><Link to="/profiles">Profiles</Link></li>
-          <li><Link to="" onClick={handleLogout}>LOG OUT</Link></li>
-          <li><Link to="/change-password">Change Password</Link></li>
-          <li><Link to="/fortunes">Fortunes</Link></li>
+          <li><Link to='/profiles'>Profiles</Link></li>
+          <li><Link to='' onClick={handleLogout}>LOG OUT</Link></li>
+          <li><Link to='/change-password'>Change Password</Link></li>
+          <li><Link to='/fortunes'>Fortunes</Link></li>
         </ul>
         :
         <ul>
-          <li><Link to="/login">Log In</Link></li>
-          <li><Link to="/signup">Sign Up</Link></li>
+          <li><Link to='/login'>Log In</Link></li>
+          <li><Link to='/signup'>Sign Up</Link></li>
         </ul>
       }
     </nav>
